fix(server): wire dead capacity config to the HTTP listener
server.workers, server.max_connections, server.timeout, and
server.max_body_size were parsed, validated, and documented but had no
runtime consumers: the listener used defaults (1x workers, no
connection cap, 5s request-head timeout) and JSON bodies were silently
capped at actix's hardcoded 2MB regardless of the configured limit.

- start(): apply worker_count(), client_request_timeout(timeout), and
  max_connections to the actix builder; timeout bounds only how long a
  client may take to deliver the request head — it never bounds
  handler or streaming duration, matching the outbound-client policy
- create_app(): register JsonConfig/FormConfig limits from
  server.max_body_size at app level and thread the value into the AI
  routes' scoped JSON configs, which previously shadowed any app-level
  limit with actix's 2MB default
- document the precise scope of timeout and max_body_size in
  ServerConfig
- move http.rs's inline test module to http_tests.rs so the core file
  stays under the 800-line guard
