+++ tests/gemini_sdk_routes.rs
@@ -32,11 +32,12 @@ mod tests {
     #[tokio::test]
     async fn gemini_sdk_routes_without_provider_fail_closed() {
         let state = build_test_state(Vec::new()).await;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -83,11 +84,12 @@ mod tests {
             ModelLimitConfig::new(100.0, ResetPeriod::Monthly),
         );
         let budget_limits = state.budget_limits.clone();
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -153,7 +155,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -190,11 +197,12 @@ mod tests {
             .unified_router
             .add_model_alias("public-gemini", "gemini@prod")
             .expect("runtime alias should install");
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -237,11 +245,12 @@ mod tests {
             ModelLimitConfig::new(100.0, ResetPeriod::Monthly),
         );
         let budget_limits = state.budget_limits.clone();
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -292,11 +301,12 @@ mod tests {
             vec!["gemini-3.1-flash-lite".to_string()],
         )])
         .await;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -327,11 +337,12 @@ mod tests {
             vec!["gemini-3.1-flash-lite".to_string()],
         )])
         .await;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -365,7 +376,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -407,7 +423,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -439,11 +460,12 @@ mod tests {
         let mock = MockGeminiServer::launch().await;
         let state =
             build_test_state(vec![gemini_provider("gemini", &mock.base_url, Vec::new())]).await;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -476,11 +498,12 @@ mod tests {
         state
             .budget_limits
             .record_spend("gemini", "gemini-3.1-flash-lite", 0.01);
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
@@ -514,11 +537,12 @@ mod tests {
             "gemini-3.1-flash-lite",
             ModelLimitConfig::new(0.000001, ResetPeriod::Monthly),
         );
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let response = test::call_service(
