+++ tests/gemini_sdk_routes/runtime_provider_tests.rs
@@ -67,11 +67,12 @@ async fn gemini_sdk_route_executes_selected_runtime_provider_snapshot() {
         "replacement-header",
     )];
     state.config.store(replaced_config);
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let response = test::call_service(
         &app,
@@ -156,11 +157,12 @@ async fn gemini_sdk_route_executes_selected_runtime_provider_snapshot() {
     let mut replaced_config = state.config().as_ref().clone();
     replaced_config.gateway.providers = vec![replacement];
     state.config.store(replaced_config);
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let started = Instant::now();
     let response = test::call_service(
@@ -184,11 +186,12 @@ async fn gemini_sdk_route_accepts_only_closed_runtime_alias_set() {
         .enumerate()
     {
         let state = build_test_state(vec![gemini_provider(name, &mock.base_url, Vec::new())]).await;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
         let response = test::call_service(
             &app,
@@ -207,11 +210,12 @@ async fn gemini_sdk_route_accepts_only_closed_runtime_alias_set() {
         .map(|name| gemini_provider(name, &mock.base_url, Vec::new()))
         .collect();
     let state = build_test_state(rejected).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let response = test::call_service(
         &app,
@@ -237,11 +241,12 @@ async fn gemini_sdk_route_accepts_native_gemini_runtime() {
     );
     provider.provider_type = "gemini".to_string();
     let state = build_test_state(vec![provider]).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let response = test::call_service(
         &app,
@@ -261,11 +266,12 @@ async fn gemini_sdk_route_accepts_native_gemini_runtime() {
     );
     unrelated.provider_type = "gemini".to_string();
     let state = build_test_state(vec![unrelated]).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let response = test::call_service(
         &app,
@@ -395,11 +401,12 @@ async fn gemini_sdk_stream_client_cancel_is_health_neutral_and_settles_observed_
         .expect("runtime deployment");
     let successes = deployment.state.success_requests.load(Ordering::Relaxed);
     let failures = deployment.state.fail_requests.load(Ordering::Relaxed);
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let response = test::call_service(
         &app,
@@ -463,11 +470,12 @@ async fn gemini_sdk_route_network_error_does_not_leak_provider_key() {
         vec!["gemini-3.1-flash-lite".to_string()],
     )])
     .await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let response = test::call_service(
@@ -495,11 +503,12 @@ async fn gemini_sdk_route_redacts_key_from_upstream_error_body() {
         vec!["gemini-3.1-flash-lite".to_string()],
     )])
     .await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let response = test::call_service(
@@ -539,11 +548,12 @@ async fn gemini_sdk_stream_route_does_not_charge_upstream_error_body() {
         ModelLimitConfig::new(100.0, ResetPeriod::Monthly),
     );
     let budget_limits = state.budget_limits.clone();
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let response = test::call_service(
@@ -606,11 +616,12 @@ async fn gemini_sdk_stream_route_settles_reserved_spend_after_output_then_read_e
         .expect("runtime deployment");
     let successes = deployment.state.success_requests.load(Ordering::Relaxed);
     let failures = deployment.state.fail_requests.load(Ordering::Relaxed);
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let request = test::TestRequest::post()
@@ -691,11 +702,12 @@ async fn gemini_sdk_stream_heartbeat_before_read_error_does_not_charge_reservati
     let (key_scope, key_context) = route_key_budget_context(&state).await;
     let budget_manager = state.budget_manager.clone();
     let budget_limits = state.budget_limits.clone();
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let request = test::TestRequest::post()
         .uri("/v1beta/models/gemini-3.1-flash-lite:streamGenerateContent")
@@ -735,11 +747,12 @@ async fn gemini_sdk_stream_body_is_not_cut_off_by_ordinary_timeout() {
     );
     provider.timeout = 1;
     let state = build_test_state(vec![provider]).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
     let started = Instant::now();
