+++ tests/image_router_fallback_routes.rs
@@ -320,7 +320,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -360,7 +365,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -400,7 +410,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -443,7 +458,12 @@ mod tests {
                     srv.call(req)
                 })
                 .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
+                .configure(|cfg| {
+                    litellm_rs::server::routes::ai::configure_routes(
+                        cfg,
+                        litellm_rs::config::models::default_max_body_size(),
+                    )
+                }),
         )
         .await;
 
@@ -484,11 +504,12 @@ mod tests {
         state
             .budget_limits
             .record_spend("openai-primary", "image-alias", 0.01);
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let resp = test::call_service(
@@ -534,11 +555,12 @@ mod tests {
             .calculate_loaded_usage_cost_for_provider("openai", "gpt-image-1-mini", &expected_usage)
             .expect("image generation pricing should be available")
             .total_cost;
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let resp = test::call_service(
@@ -586,11 +608,12 @@ mod tests {
             ProviderLimitConfig::new(100.0, ResetPeriod::Monthly),
         );
         let budget_limits = state.budget_limits.clone();
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let resp = test::call_service(
@@ -637,11 +660,12 @@ mod tests {
             ProviderLimitConfig::new(100.0, ResetPeriod::Monthly),
         );
         let budget_limits = state.budget_limits.clone();
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let resp = test::call_service(
@@ -691,11 +715,12 @@ mod tests {
             ProviderLimitConfig::new(100.0, ResetPeriod::Monthly),
         );
         let budget_limits = state.budget_limits.clone();
-        let app = test::init_service(
-            App::new()
-                .app_data(web::Data::new(state))
-                .configure(litellm_rs::server::routes::ai::configure_routes),
-        )
+        let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+            litellm_rs::server::routes::ai::configure_routes(
+                cfg,
+                litellm_rs::config::models::default_max_body_size(),
+            )
+        }))
         .await;
 
         let resp = test::call_service(
