+++ tests/integration/http_routes_tests.rs
@@ -214,7 +214,12 @@ mod tests {
             .app_data(budget_limits)
             .wrap(AuthMiddleware)
             .configure(routes::health::configure_routes)
-            .configure(routes::ai::configure_routes)
+            .configure(|cfg| {
+                litellm_rs::server::routes::ai::configure_routes(
+                    cfg,
+                    litellm_rs::config::models::default_max_body_size(),
+                )
+            })
     }
 
     // ---------------------------------------------------------------
