+++ tests/tests/fine_tuning_routes_policy_tests.rs
@@ -9,11 +9,12 @@ async fn public_only_fine_tuning_route_rejects_loopback_before_connect() {
     let mut provider = fine_tuning_provider(&format!("http://{address}/v1"));
     provider.endpoint_access = ProviderEndpointAccess::PublicOnly;
     let state = build_test_state(vec![provider]).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let response = test::call_service(
