+++ tests/tests/rerank_routes_policy_tests.rs
@@ -3,11 +3,12 @@ use super::*;
 async fn assert_public_only_rerank_rejects_loopback(mut provider: ProviderConfig, request: Value) {
     provider.endpoint_access = ProviderEndpointAccess::PublicOnly;
     let state = build_test_app_state(vec![provider]).await;
-    let app = test::init_service(
-        App::new()
-            .app_data(web::Data::new(state))
-            .configure(litellm_rs::server::routes::ai::configure_routes),
-    )
+    let app = test::init_service(App::new().app_data(web::Data::new(state)).configure(|cfg| {
+        litellm_rs::server::routes::ai::configure_routes(
+            cfg,
+            litellm_rs::config::models::default_max_body_size(),
+        )
+    }))
     .await;
 
     let response = test::call_service(
